Adding Support for XBox One Controller
This commit enables the use of the XBox One controller with pacmod_game_control.
The XBox One controller uses the same button layout as the Logitech F310, which simplifies
the code. Additionally, it should be noted that the XBox controller must be plugged
into the computer using a USB to Micro USB cable.