Sending shift command on enable/disable of system
Before this commit, PACMod game control would not send a command to the shifter
system when the enable or disable button was pressed. As such, the shifter module
did not register the enable/disable change until it received the next command
with the enable bit set. After this commit, PACMod game control sends a
command with the new enable/disable status when the button is pressed. The
gear value sent with this message is the same as the last reported gear.
Originally, I wanted to send an empty messge, but this would have resulted in a
shift to Park.