Added enable/disable logic for turn signal command
Before this commit, the turn signal was latching to "enable" and
would not disable in a similar manner to the previous behavior
of the shift command. After this commit, the enable/disable button
on the controller will send a turn signal command with the enable flag
set appropriately, and populated with the last command read so as not
to overwrite the command state of the turn signal.