Adding publish control factory to return correct publish control version
Before this commit, the publish control class for different boards had to be explicitly
instantiated, resulting in one instance per board type needing to be instantiated.
Currently, that would only be 2, but in the future, it could be many more. After this
commit, the factory can be used to return an instance of any existing and future board
types.