Adding message send on enable/disable logic
Before this commit, pacmod_game_control would only send new messages when enabled,
and would not send them ON enable or disable. After this commit, the node will
send messages with the appropriate flag when the system is enabled OR when the
enable state changes, either from disabled to enabled, or enabled to disabled.