Adding shift rpt and turn rpt subscribers for PM3
Before this commit, pressing the enable/disable buttons on the
controller would not send a turn command nor shift command with
the enable flag set appropriately, resulting in the turn or shift
system remaining enabled or disabled until the next regular command
was sent. After this commit, hitting the enable or disable sends
both turn and shift commands with the new flag set correctly.