Adding Major Board Rev as Selector, unique_ptr, and board exception
Before this commit, the PACMod board version was selected via a boolean in the
launch file, and the factory class provided the publisher based on this boolean.
Additionally, there was the potential for a memory leak when the factory class
returned a raw pointer to the new publisher. After this commit, the PACMod
board revision is denoted in the launch file as an integer (the major rev).
This integer is used to evaluate booleans in the launch file, which then determine
which PACMod driver version to launch. The major revision number is also used by
the factory, allowing for extensibility, and the factory will also throw an
invalid argument exception if an invalid board number is set. Finally, the raw
pointer to the publisher has been replaced with a unique_ptr for memory
management.