Revising launch file to maintain support for ROS Indigo
Before this commit, the launch file made use of "eval" for arguments
to create a boolean by comparing against the "pacmod_board_rev" argument.
This boolean was used to launch the appropriate PACMod driver. After this
commit, the board rev parameter still exists to be sent to the game control
node/factory, but there is an additional "is_pacmod_3" flag that needs to
be set to launch the correct PACMod driver. ROS Indigo does not have support
for "eval" in launch files.