Adding send clear override on first message after enable
Before this commit, the clear_override flag was not set on any
message being sent by PACMod game control, as it did not exist.
After this commit, the game control node will set the clear_override
flag to true on the first message after enable on each of the
PACMod system messages.