Fixing threading problem with pacmod_enable.
The pacmod_enable variable can be modified in two separate threads:
the callback for joy (in the case of a user-initiated enable/disable)
or the callback for the PACMod's enable status topic (in case of an
override or other disable). This necessitates having a "local" copy
of the enable state through the publishing process to keep from having
to lock/unlock a mutex every time we need the current enable/disable
state. This commit converts the "local_enable" variable to one that
is attached to PublishControl and is only updated on a joy callback.